move type into types folder
